feat(歌单与歌曲): :sparkles: 完成歌单与歌曲聚合根的持久化仓库

+++ apps/music-api/music-main/src/main/resources/db/V1.3.0__Create_List_song_Table.sql
@@ -2,8 +2,8 @@ DROP TABLE IF EXISTS `list_song`;
 CREATE TABLE `list_song`
 (
     `id`           BIGINT  NOT NULL,
-    `song_id`      int(10) unsigned NOT NULL,
-    `song_list_id` int(10) unsigned NOT NULL,
+    `song_id`      BIGINT  NOT NULL,
+    `song_list_id` BIGINT  NOT NULL,
     PRIMARY KEY (`id`),
     INDEX `song_list_idx` (song_id, song_list_id)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8 COMMENT='歌单歌曲关联表';
